Feature/monorepo (#1)
* feat: HTML 파일 삭제 및 pnpm 작업공간 구성 추가

- index.html, index.basic.html, index.advanced.html 파일 삭제
- pnpm-workspace.yaml 파일 추가하여 패키지 관리 및 작업공간 구성
- package.json에서 스크립트 및 의존성 업데이트로 개발 환경 개선
- tsconfig.json 및 vite.config.js 파일 삭제로 불필요한 설정 정리

✅ 프로젝트 구조 간소화 및 패키지 관리 효율성 향상

* feat: 기본 애플리케이션 구조 및 컴포넌트 추가

- index.html, jsconfig.json, package.json, vite.config.js 파일 추가로 기본 애플리케이션 구조 설정
- src 디렉토리에 app.js, main.js, setupTests.js 파일 추가하여 초기화 및 테스트 환경 구성
- CartDisplay, CartEventHandler, CartItem, HelpModal, NotificationBar, OrderSummary, ProductSelector, StockInfo, DiscountEngine, PointsCalculator, PriceCalculator, StockCalculator 등 다양한 컴포넌트 및 헬퍼 모듈 구현
- 각 컴포넌트에 대한 단위 테스트 추가로 코드 안정성 및 품질 향상

✅ 새로운 애플리케이션 구조로 유지보수성 및 확장성 향상

* feat: 쇼핑 카트 애플리케이션 초기화 및 기본 구조 추가

- index.html, package.json, vite.config.js, main.js, setupTests.js 파일 추가로 기본 애플리케이션 구조 설정
- 장바구니 기능을 위한 다양한 컴포넌트 및 로직 구현
- 기본적인 상품 목록 및 장바구니 기능을 위한 초기화 로직 추가
- 단위 테스트를 위한 기본 테스트 환경 구성

✅ 새로운 애플리케이션 구조로 기능 확장성 및 유지보수성 향상

* feat: 고급 쇼핑 카트 애플리케이션 초기화 및 기본 설정 추가

- index.html, package.json, tsconfig.json, vite.config.js 파일 추가로 고급 애플리케이션 구조 설정
- Tailwind CSS를 활용한 기본 스타일 및 레이아웃 구성
- Vite를 이용한 개발 및 빌드 스크립트 설정
- TypeScript 설정으로 코드 품질 및 안정성 향상

✅ 새로운 고급 애플리케이션 구조로 기능 확장성 및 유지보수성 향상
