+++ apps/advanced/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@clean-code/advanced",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "vitest"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "packageManager": "pnpm@10.13.1",
+  "dependencies": {
+    "@clean-code/eslint": "workspace:*",
+    "vite": "^7.0.5",
+    "vitest": "^3.2.4"
+  }
+}
