+++ apps/advanced/vite.config.js
