+++ apps/basic/src/app.js
@@ -1,10 +1,10 @@
-import { DOMElementManager } from './core/DomManager.js';
-import { EventManager } from './core/EventManager.js';
-import { UIUpdater } from './core/UIUpdater.js';
-import { CalculationEngine } from './helpers/CalculationEngine.js';
-import { PromotionManager } from './helpers/PromotionManager.js';
-import { ShoppingCartState } from './helpers/ShoppingCartState.js';
-import { ApplicationService } from './services/ApplicationService.js';
+import { DOMElementManager } from '@core/DomManager';
+import { EventManager } from '@core/EventManager';
+import { UIUpdater } from '@core/UIUpdater';
+import { CalculationEngine } from '@helpers/CalculationEngine';
+import { PromotionManager } from '@helpers/PromotionManager';
+import { ShoppingCartState } from '@helpers/ShoppingCartState';
+import { ApplicationService } from '@services/ApplicationService';
 
 export class BootstrapApplication {
   constructor() {
