+++ apps/basic/src/constants/DiscountPolicies.js
