+++ apps/basic/src/core/EventManager.js
