+++ apps/basic/src/helpers/PointsCalculator.js
