+++ apps/basic/src/helpers/ShoppingCartState.js
