+++ apps/basic/src/helpers/StockCalculator.js
