+++ apps/original/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@clean-code/original",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "vitest"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "packageManager": "pnpm@10.13.1",
+  "dependencies": {
+    "@clean-code/eslint": "workspace:*",
+    "vite": "catalog:vite",
+    "vitest": "catalog:test",
+    "jsdom": "catalog:test",
+    "@testing-library/jest-dom": "catalog:test",
+    "@testing-library/user-event": "catalog:test"
+  }
+}
